requirements.txt
Flask==3.0.0; python_version > '3.6'
Flask==2.3.3; python_version < '3.7'
Werkzeug==3.0.1; python_version > '3.6'
Werkzeug==2.3.7; python_version < '3.7'
gunicorn==20.1.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.25.1